# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Соседи API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/me/profile/portfolio": {
      "get": {
        "tags": [
          "specialists",
          "specialists"
        ],
        "summary": "Get My Portfolio",
        "description": "Работы своего профиля по порядку (S37); 404 `profile_not_found` — профиля нет.",
        "operationId": "specialists_get_my_portfolio",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка в формате RFC 9457",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "specialists",
          "specialists"
        ],
        "summary": "Add My Work",
        "description": "Работа из загруженного файла — в конец; лимит — 409 `portfolio_full` (kind, limit).",
        "operationId": "specialists_add_my_work",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка в формате RFC 9457",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Ключ операции: повтор с тем же ключом вернёт тот же ответ",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ]
      }
    },
    "/api/v1/me/profile/portfolio/order": {
      "put": {
        "tags": [
          "specialists",
          "specialists"
        ],
        "summary": "Reorder My Portfolio",
        "description": "Новый порядок — все работы; не те или не все — 422 `invalid_portfolio`.",
        "operationId": "specialists_reorder_my_portfolio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioOrderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка в формате RFC 9457",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/profile/portfolio/{item_id}": {
      "patch": {
        "tags": [
          "specialists",
          "specialists"
        ],
        "summary": "Caption My Work",
        "description": "Подпись работы; пустая — без подписи.",
        "operationId": "specialists_caption_my_work",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkCaptionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка в формате RFC 9457",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "specialists",
          "specialists"
        ],
        "summary": "Remove My Work",
        "description": "Убрать работу; её файл удаляется.",
        "operationId": "specialists_remove_my_work",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "description": "Ошибка в формате RFC 9457",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldErrorOut": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldErrorOut"
      },
...
      "MediaRefOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/MediaVariantOut"
            },
            "type": "array",
            "title": "Variants"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "placeholder",
          "variants",
          "video_url",
          "duration_ms"
        ],
        "title": "MediaRefOut",
        "description": "Файл работы или фото профиля: пока он обрабатывается, вариантов нет."
      },
...
      "MediaVariantOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "width",
          "height"
        ],
        "title": "MediaVariantOut"
      },
...
      "PortfolioLimitsOut": {
        "properties": {
          "image": {
            "type": "integer",
            "title": "Image",
            "default": 60
          },
          "video": {
            "type": "integer",
            "title": "Video",
            "default": 6
          }
        },
        "type": "object",
        "title": "PortfolioLimitsOut"
      },
      "PortfolioOrderIn": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 66,
            "minItems": 1,
            "title": "Item Ids"
          }
        },
        "type": "object",
        "required": [
          "item_ids"
        ],
        "title": "PortfolioOrderIn"
      },
      "PortfolioOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkOut"
            },
            "type": "array",
            "title": "Items"
          },
          "limits": {
            "$ref": "#/components/schemas/PortfolioLimitsOut"
          }
        },
        "type": "object",
        "required": [
          "items",
          "limits"
        ],
        "title": "PortfolioOut",
        "description": "Портфолио в кабинете S37: работы по порядку и лимиты (60 фото, 6 роликов)."
      },
...
      "ProblemOut": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "restriction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restriction"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "min_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Version"
          },
          "documents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents"
          },
          "document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "trace_id"
        ],
        "title": "ProblemOut",
        "description": "Ошибка RFC 9457. Клиент ветвится по `code`, текст для человека — `detail`."
      },
...
      "WorkCaptionIn": {
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "caption"
        ],
        "title": "WorkCaptionIn"
      },
      "WorkIn": {
        "properties": {
          "media_id": {
            "type": "string",
            "format": "uuid",
            "title": "Media Id"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "media_id"
        ],
        "title": "WorkIn"
      },
      "WorkKind": {
        "type": "string",
        "enum": [
          "image",
          "video"
        ],
        "title": "WorkKind"
      },
...
      "WorkOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkKind"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaRefOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "caption",
          "position",
          "media"
        ],
        "title": "WorkOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access JWT из POST /auth/telegram",
        "scheme": "bearer"
      }
    }
  }
}