# packages/api-client/src/generated/model/mediaVariantOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Соседи API
 * OpenAPI spec version: 1.0.0
 */

export interface MediaVariantOut {
  name: string;
  url: string;
  width: number;
  height: number;
}