# packages/api-client/src/generated/model/portfolioOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Соседи API
 * OpenAPI spec version: 1.0.0
 */
import type { PortfolioLimitsOut } from './portfolioLimitsOut.ts';
import type { WorkOut } from './workOut.ts';

/**
 * Портфолио в кабинете S37: работы по порядку и лимиты (60 фото, 6 роликов).
 */
export interface PortfolioOut {
  items: WorkOut[];
  limits: PortfolioLimitsOut;
}