# packages/api-client/src/generated/model/workKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Соседи API
 * OpenAPI spec version: 1.0.0
 */

export type WorkKind = (typeof WorkKind)[keyof typeof WorkKind];

export const WorkKind = {
  image: 'image',
  video: 'video',
} as const;