# packages/api-client/src/generated/model/workOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Соседи API
 * OpenAPI spec version: 1.0.0
 */
import type { MediaRefOut } from './mediaRefOut.ts';
import type { WorkKind } from './workKind.ts';

export interface WorkOut {
  id: string;
  kind: WorkKind;
  caption: string | null;
  position: number;
  media: MediaRefOut | null;
}